Added node installation conditional
Added node installation conditional to first check if Node is already installed in `$build_dir/vendor/node`. If it is, then use that installation. Otherwise, download and install Node as normal.

This helps circumvent problems when building an app that already has Node installed in `$build_dir/vendor/node` from a previous build with this buildpack.